+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 
 ## Master
 
+- Add workaround for Xcode 12 [@f-meloni][] - [#372](https://github.com/danger/swift/pull/3672)
+- Deploy Danger swift Docker image on GitHub Container Registry [@f-meloni][] - [#367](https://github.com/danger/swift/pull/367)
 - Create sync utils function to make async calls sync [@f-meloni][] - [#369](https://github.com/danger/swift/pull/369)
 
 ## 3.4.1