+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 
 ## Master
 
+- Make `draft` optional for some repos which Draft pull requests are not available in GitHub Pulls [@417-72KI][] - [#378](https://github.com/danger/swift/pull/378)
 - Add Link Relations for GitHub PR [@417-72KI][] - [#368](https://github.com/danger/swift/pull/368)
 
 ## 3.5.0